ufork.js
This commit extracts reusable code from the debugger into its own module,
so the uFork WASM can be used by generic applications.
